Handle exceptions raised from Requests
This particular approach allows us to avoid having to rewrite every unit
test that makes an assertion about the mocked session having its get,
post, patch, etc. methods called. It is sub-optimal, but until we have
the time and patience to rewrite the tests, thi swill have to make do.

Closes #597