Add Organization.all_events.
`username` is currrently a required argument for `all_events` because GitHub's
API endpoint requires the username of the authenticated user. It is possible
that this keyword argument will not be necessary in the future, thus I added it
to the end of the parameter list in order to minimize its use as a positional
argument.